changes to enable the addition of different messengers
The script will now spawn a child process depending of the selected messenger to start the script. Because of how the script was layed out, messengers will temporarily have its own modules even when they are just slightly different from the ones another messenger is using. It was never planned to add new messengers to this project, which is why everything is designed extremely strictly for Telegram. Hopefully, I can loosen things up a bit in later updates.